Capacity Intro Edits
Added "may wish" because it sounded to me too much like these are the only possible/valuable; rather they are the only ones we can actually provide at the moment. Dropped 's' off deployments, because they are per deployment and refer to it as deployment singular on KPI page. Removed the one sentence because copy should be reduced ideally and it wasn't that valuable of a sentence.