qualify Rolling App Deployments GA
customers are getting confused by the fact that rolling app deployments are GA in TAS v2.7, but the CLI doesn't support them yet.
I've updated the content to reflect the fact that while the feature was GA'd from an API and Apps Man perspective, the v6 CLI that's compatible with TAS v2.7 only has an experimental implementation of the rolling deployment feature.
The feature is fully GA within the cf CLI as of cf CLI v7, which is compatible with TAS v2.10 and above (it cannot be used against TAS 2.7-2.9).

Please feel free to change the copy as you wish. 
The main point I need customers to understand is that the feature is only GA'd from a backend and AppsMan perspective and the CLI's GA of the feature is later in TAS v2.10 and via the V7 cf CLI. feel free to reach out to me via slack if you need to clarify anything. Thanks so much!